Learney 0.1.0 - I'm so sorry
This commit is the largest I've ever done. I'm not proud of that fact. I
promise to do better next time. The shame I feel from this shall
reverberate through my bones and those of my descendents for millenia.

Onto the code: this contains the slack question bot (yet to be tested on
staging), introduces testing, black and linting so now there's some
actual code standards.