Fix Slackbot
Found a bug whereby learned_concepts and goal_concepts were stored as
strings instead of as dictionaries. This led to issues where the
Slackbot was expecting a dict. I fixed it and changed all the live data
to dicts, but it's concerning this is happening - it's a result of a
lack of testing.