Add session_id entry to user action tables
Also tracks when the last action performed was, stores this in the
session object and then when a new page is loaded if this was more than
an hour ago, the session key is updated.

Sessions are all modified, so all will now be saved in the Sessions
django builtin Sessions table.