Fix HTTP_HOST ALLOWED_HOSTS warning
For some reason, on prod, it requires a token and on staging it does not.
This bamboozled me and led to Sentry being overrun by issues for a long
time, to the point that Sentry stopped recording our issues. I've been
papering over the cracks of this for a very long time and I think it's
starting to bite me - there are lots of errors now being recorded on Sentry
(reactivated as I started the free trial) that are affecting people every
MINUTE on the site