Frontend id -> params & template id and reverse utils
These are useful because we want unique ids for each question in the
frontend, but this data is stored and generated separately in the
backend. Removing the questionModel class removed one problem - if we
tweak a questionTemplate then how do we translate that change across
to all the existing questionModels associated with it? Now the specific
instances of each question are generated on the fly, so this step isn't
necessary.