Fix streak bug!
The bug was that the function checking whether a question batch
was finished yesterday didn't work because it looked at whether
the most recent batch was completed yesterday, not whether one
was completed yesterday at all. This led to incorrect reporting
when you have a streak and keep it up today and reload the page.